Handle hostnames that start with numbers.
This involved needing to look at more than one token at a time, so
this patch moves from read() to mmap() of files, as a way to gracefully
rewind.

See RT ticket #16516 for (a lot) more.

+++ includes/dhcpd.h
@@ -52,6 +52,7 @@
 #include <string.h>
 #include <stdlib.h>
 #include <sys/stat.h>
+#include <sys/mman.h>
 #include <ctype.h>
 #include <time.h>
 
@@ -272,8 +273,8 @@ struct parse {
 	int warnings_occurred;
 	int file;
 	char *inbuf;
-	unsigned bufix, buflen;
-	unsigned bufsiz;
+	size_t bufix, buflen;
+	size_t bufsiz;
 };
 
 /* Variable-length array of data. */
